+++ FrontEndScripts/typings/showdown.d.ts
@@ -1,0 +1,11 @@
+﻿declare module showdown {
+
+  export class Converter {
+
+    constructor(options?: any);
+
+    public makeHtml(value: string): string;
+
+  }
+
+}